+++ src/ru/loghorrean/veganShop/models/database/entities/GeneralDishInOrder.java
@@ -3,9 +3,9 @@ package ru.loghorrean.veganShop.models.database.entities;
 public class GeneralDishInOrder extends DatabaseEntity {
     private GeneralDish dish;
     private Order order;
-    private int amount;
+    private float amount;
 
-    public GeneralDishInOrder(int id, GeneralDish dish, Order order, int amount) {
+    public GeneralDishInOrder(int id, GeneralDish dish, Order order, float amount) {
         super(id);
         this.dish = dish;
         dish.addOrder(order);
@@ -14,7 +14,7 @@ public class GeneralDishInOrder extends DatabaseEntity {
         this.amount = amount;
     }
 
-    public GeneralDishInOrder(GeneralDish dish, Order order, int amount) {
+    public GeneralDishInOrder(GeneralDish dish, Order order, float amount) {
         this(-1, dish, order, amount);
     }
 
@@ -38,11 +38,11 @@ public class GeneralDishInOrder extends DatabaseEntity {
         this.order.addGeneralDish(dish);
     }
 
-    public int getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
